requirements.txt
asgiref==3.3.1
bcrypt==3.2.0
cffi==1.14.4
Django==3.1.4
django-cors-headers==3.6.0
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
mysqlclient==2.0.2
pycparser==2.20
PyJWT==1.7.1
pytz==2020.4
six==1.15.0
sqlparse==0.4.1
